Refactor game state provider into a new file
